+++ instrumentation/opentelemetry-instrumentation-elasticsearch/tests/test_elasticsearch.py
@@ -11,6 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
 import json
 import os
 import threading
@@ -24,15 +25,17 @@ from elasticsearch_dsl import Search
 
 import opentelemetry.instrumentation.elasticsearch
 from opentelemetry import trace
-from opentelemetry.instrumentation.elasticsearch import ElasticsearchInstrumentor
+from opentelemetry.instrumentation.elasticsearch import (
+    ElasticsearchInstrumentor,
+)
 from opentelemetry.instrumentation.elasticsearch.utils import sanitize_body
 from opentelemetry.semconv.trace import SpanAttributes
 from opentelemetry.test.test_base import TestBase
 from opentelemetry.trace import StatusCode
 
-major_version = elasticsearch.VERSION[0]
-
 from . import sanitization_queries as queries
+
+major_version = elasticsearch.VERSION[0]
 
 if major_version == 7:
     from . import helpers_es7 as helpers  # pylint: disable=no-name-in-module
@@ -69,7 +72,7 @@ class TestElasticsearchIntegration(TestBase):
     def setUp(self):
         super().setUp()
         self.tracer = self.tracer_provider.get_tracer(__name__)
-        ElasticsearchInstrumentor().instrument(sanitize_query=False)
+        ElasticsearchInstrumentor().instrument()
 
     def tearDown(self):
         super().tearDown()
@@ -156,7 +159,9 @@ class TestElasticsearchIntegration(TestBase):
         self.assertEqual(1, len(spans))
         self.assertEqual(spans[0].name, "Elasticsearch/test-index/_doc/:id")
         self.assertEqual("False", spans[0].attributes["elasticsearch.found"])
-        self.assertEqual("True", spans[0].attributes["elasticsearch.timed_out"])
+        self.assertEqual(
+            "True", spans[0].attributes["elasticsearch.timed_out"]
+        )
         self.assertEqual("7", spans[0].attributes["elasticsearch.took"])
 
     def test_trace_error_unknown(self, request_mock):
@@ -183,7 +188,9 @@ class TestElasticsearchIntegration(TestBase):
         span = spans[0]
         self.assertFalse(span.status.is_ok)
         self.assertEqual(span.status.status_code, code)
-        self.assertEqual(span.status.description, f"{type(exc).__name__}: {exc}")
+        self.assertEqual(
+            span.status.description, f"{type(exc).__name__}: {exc}"
+        )
 
     def test_parent(self, request_mock):
         request_mock.return_value = (1, {}, {})
@@ -259,12 +266,14 @@ class TestElasticsearchIntegration(TestBase):
     def test_dsl_search_sanitized(self, request_mock):
         # Reset instrumentation to use sanitized query (default)
         ElasticsearchInstrumentor().uninstrument()
-        ElasticsearchInstrumentor().instrument()
+        ElasticsearchInstrumentor().instrument(sanitize_query=True)
 
         # update expected attributes to match sanitized query
         sanitized_search_attributes = self.search_attributes.copy()
         sanitized_search_attributes.update(
-            {SpanAttributes.DB_STATEMENT: "{'query': {'bool': {'filter': '?'}}}"}
+            {
+                SpanAttributes.DB_STATEMENT: "{'query': {'bool': {'filter': '?'}}}"
+            }
         )
 
         request_mock.return_value = (1, {}, '{"hits": {"hits": []}}')
@@ -278,7 +287,6 @@ class TestElasticsearchIntegration(TestBase):
         self.assertEqual(1, len(spans))
         self.assertEqual(span.name, "Elasticsearch/<target>/_search")
         self.assertIsNotNone(span.end_time)
-        print("~~~~~~~~~~", span.attributes)
         self.assertEqual(
             span.attributes,
             sanitized_search_attributes,
@@ -392,7 +400,9 @@ class TestElasticsearchIntegration(TestBase):
         spans = self.get_finished_spans()
 
         self.assertEqual(1, len(spans))
-        self.assertEqual("GET", spans[0].attributes[request_hook_method_attribute])
+        self.assertEqual(
+            "GET", spans[0].attributes[request_hook_method_attribute]
+        )
         self.assertEqual(
             f"/{index}/_doc/{doc_id}",
             spans[0].attributes[request_hook_url_attribute],
@@ -407,7 +417,9 @@ class TestElasticsearchIntegration(TestBase):
 
         def response_hook(span, response):
             if span and span.is_recording():
-                span.set_attribute(response_attribute_name, json.dumps(response))
+                span.set_attribute(
+                    response_attribute_name, json.dumps(response)
+                )
 
         ElasticsearchInstrumentor().uninstrument()
         ElasticsearchInstrumentor().instrument(response_hook=response_hook)
@@ -454,11 +466,14 @@ class TestElasticsearchIntegration(TestBase):
 
     def test_body_sanitization(self, _):
         self.assertEqual(
-            sanitize_body(queries.interval_query), str(queries.interval_query_sanitized)
+            sanitize_body(queries.interval_query),
+            str(queries.interval_query_sanitized),
         )
         self.assertEqual(
-            sanitize_body(queries.match_query), str(queries.match_query_sanitized)
+            sanitize_body(queries.match_query),
+            str(queries.match_query_sanitized),
         )
         self.assertEqual(
-            sanitize_body(queries.filter_query), str(queries.filter_query_sanitized)
+            sanitize_body(queries.filter_query),
+            str(queries.filter_query_sanitized),
         )