+++ CHANGELOG.md
@@ -9,8 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add default query sanitization for elasticsearch db.statement attribute
-  ([#1545](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/1545))
+- Add optional query sanitization for elasticsearch db.statement attribute
+  ([#1598](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/1598))
 - `opentelemetry-instrumentation-redis` Add `sanitize_query` config option to allow query sanitization.
   ([#1572](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/1572))
 - `opentelemetry-instrumentation-celery` Record exceptions as events on the span.